Update to use Vapoursynth v4 API (#30)
In r73, the Vapoursynth developers removed the v3 VSScript API. Therefore, in order to unbreak the Rust Vapoursynth ecosystem, which was still using it, this PR moves the crate onto the v4 Vapoursynth API. It is worth noting that there is another "vapoursynth4" crate on crates.io, but it appears to have never been finished, and it doesn't follow idiomatic and safe practices for its public interface, so we should prefer updating this crate. It was not feature complete enough to work with av1an.

The intention was to maintain as much of the public API interface (vapoursynth-rs crate) as possible, though inevitably there were a few breaking changes. The vapoursynth-sys crate has been updated to be auto-generated from the v4 Vapoursynth C header files, to ensure full correctness and allow for easy future updates. I manually tested the new version against av1an, and it appears to pass all of its tests.

It is also noteworthy that Vapoursynth appears to have removed i686 builds from Github. The CI files have been adjusted not to run on this platform, as they are only officially supporting x64 now.